+++ src/main/java/faang/school/accountservice/dto/balance/BalanceUpdateRequest.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class BalanceUpdateRequest {
-    private Long authorisationBalance;
-    private Long actualBalance;
+    private BigDecimal authorisationBalance;
+    private BigDecimal actualBalance;
 }
